chore: 移动常量到 constant 文件，修改文件名，导出方法统一到 index 中
