+++ packages/s2-core/src/common/constant/copy.ts
@@ -3,3 +3,6 @@ export enum CopyType {
   COL,
   ROW,
 }
+
+export const NewLine = '\r\n';
+export const NewTab = '\t';
