+++ packages/s2-core/src/utils/export/copy/pivot-data-cell-copy.ts
